STAT interrupt, fix LCD bug, fix Timer bug
Bunch of stuff to get Dr.Mario working. Found a big bugs with the LCD
when it gets tuned off.
Also tried a few mooneye test roms as well, but they seem to test more
obscure behaviours than Blargg so idk how much I'll test with them (also
they are a bit of a pain to use manually)